requirements.txt
telebot
requests
beautifulsoup4
python-dotenv